EC-7 : Now scrapes all courses on ListOfCoursesToScrape.txt.
Had to remove courses that were outdated and not offered anymore from list. Also removed a 1 credit course SOEN 491, because it was formatted wrong, should not be a problem.

Currently outputs to a text file, but the format is arranged to make it easy to store in tables once I figure that out.

Changed so that you just run scrapper once now, extract is a function that scrapper calls.

What is missing:
 - Differentiating pre-reqs from co-reqs, but we can use this for now
 - Some pre-reqs are cegep level courses, so we'll have to add those to the DB/student record somehow, or remove them manually perhaps after.
 - auto-storing in database